Changes to all files
-All buffer counters and sleeps replace with chrono package based timers
dependant on a given frequency
-New object classifer implemented (could be improved still but more
relable than previously)
-Image handlers no loner use colormap and use imagesc() instead which
distributes the scale of color properly.
-MATLAB code done without for loops, only array and matrix operations so
its a lot more efficient.
-Added point cloud plot (based off harry's x generated data and a
generated y. The plot demonstrates we need point cloud data as the
dimensions of point cloud objects are not constant (they change with
distance to the camera as we thought)
-Various stylistic changes throughout all files
-Framework set up for sending vertex data however found out that the
data in data.vertex is never populated hence leading to a segfault
whenever you try to read from it. All sub-sampling and sending work
confirmed working with constant values (just can't get the point cloud
data to read, maybe it needs an initialiser or for the node to be
registered like on softkinetic?).